Add EndpointResqlComponent
